03/04/2017
For the upcoming weeks till the end of this semester, I'll be uploading
less and less codes

Improvement for coding isn't based completely on speed. I may be getting
better at all the problems I enjoyed doing, which are the simple ones, I
could barely come up with any solutions for the hard ones.

In other words, I'll be looking through library of algorithms and data
structures and other techniques of improvement and mastering them all
before I do more programming challenges. This also means I'll be
building my own library of codes, such as all sorts of binary search
trees + trie, and the frequency of this github commit will be really
low.